Move handler to balance command module
