+++ src/commands/balance.rs
@@ -1,5 +1,14 @@
+use std::collections::HashMap;
+use std::error::Error;
+use std::fs;
+
 use chrono::NaiveDate;
-use clap::{value_parser, Arg, Command};
+use clap:: { ArgMatches, Arg, Command, value_parser };
+use prettytable::format;
+use serde_json;
+
+use crate::journal;
+use crate::report;
 
 pub fn balance_command() -> Command {
     Command::new("balance")
@@ -18,21 +27,40 @@ pub fn balance_command() -> Command {
         )
 }
 
-// #[derive(Parser)]
-// #[command(author, version, about, long_about = None)] // Read from `Cargo.toml`
-// pub struct Cli {
-//     #[command(subcommand)]
-//     pub command: Commands,
-// }
+pub fn balance_handler(report_args: &ArgMatches) {
+    let filepath_option = report_args.get_one::<String>("filepath").expect("required");
+    let input_journal: journal::Journal = load_journal(filepath_option).unwrap();
 
-// #[derive(Subcommand)]
-// pub enum Commands {
-//     Balance {
-//         #[arg(short = 'p', long, required = true)]
-//         filepath: Option<String>,
-//         #[arg(short = 'f', long)]
-//         from_date: Option<NaiveDate>,
-//         #[arg(short = 't', long)]
-//         to_date: Option<NaiveDate>,
-//     },
-// }
+    input_journal.validate();
+
+    let from_date = NaiveDate::from_ymd_opt(2000, 1, 1).unwrap();
+    let to_date = NaiveDate::from_ymd_opt(2200, 1, 1).unwrap();
+
+    let balances = report::balance::balance(&input_journal, &from_date, &to_date);
+
+    display_balances(balances);
+}
+
+fn load_journal(filepath: &String) -> Result<journal::Journal, Box<dyn Error>> {
+    let ledgerfile: String = fs::read_to_string(filepath)?.parse()?;
+    let input_journal: journal::Journal = serde_json::from_str(&ledgerfile)?;
+    Ok(input_journal)
+}
+
+fn display_balances(balances: HashMap<String, f64>) {
+    let mut account_names: Vec<&String> = balances.keys().collect();
+    account_names.sort();
+
+    let mut table = prettytable::Table::new();
+    table.set_titles(row!["Account", "Balance"]);
+
+    for account_name in account_names {
+        table.add_row(row![
+            account_name.clone(),
+            balances.get(account_name).unwrap().to_string()
+        ]);
+    }
+
+    table.set_format(*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
+    table.printstd();
+}
